fix(futurizable): move interface to be compositional interface and fix Either
