+++ src/futurizable/futurizable.ts
@@ -1,4 +1,4 @@
-import { Future } from './future';
+import { Future } from '../future/future';
 
 export interface Futurizable<T> {
   toFuture(): Future<T>;
