+++ src/futurizable/index.ts
@@ -1,0 +1,1 @@
+export { Futurizable } from './futurizable';
